requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
django-environ==0.4.5
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
pdoc3==0.9.2
pytz==2021.1
sqlparse==0.4.1
